backend/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
